No path to certificates if ssl is not enabled

+++ ad2web/certificate/views.py
@@ -25,6 +25,9 @@ def certificate_context_processor():
 @certificate.route('/', methods=['GET', 'POST'])
 @login_required
 def index():
+    use_ssl = Setting.get_by_name('use_ssl', default=False).value
+    if use_ssl == False:
+        abort(404)
     certificates = Certificate.query.all()
     ca_cert = Certificate.query.filter_by(type=CA).first()
     return render_template('certificate/index.html', certificates=certificates, ca_cert=ca_cert, active='certificates')
@@ -32,6 +35,10 @@ def index():
 @certificate.route('/generate', methods=['GET', 'POST'])
 @login_required
 def generate():
+    use_ssl = Setting.get_by_name('use_ssl', default=False).value
+    if use_ssl == False:
+        abort(404)
+
     form = GenerateCertificateForm(next=request.args.get('next'))
 
     if form.validate_on_submit():
@@ -61,6 +68,10 @@ def generate():
 @certificate.route('/<int:certificate_id>')
 @login_required
 def view(certificate_id):
+    use_ssl = Setting.get_by_name('use_ssl', default=False).value
+    if use_ssl == False:
+        abort(404)
+
     cert = Certificate.get_by_id(certificate_id)
     ca_cert = Certificate.query.filter_by(type=CA).first()
     if cert.user != current_user and not current_user.is_admin():
@@ -71,6 +82,10 @@ def view(certificate_id):
 @certificate.route('/<int:certificate_id>/download/<download_type>')
 @login_required
 def download(certificate_id, download_type):
+    use_ssl = Setting.get_by_name('use_ssl', default=False).value
+    if use_ssl == False:
+        abort(404)
+
     cert = Certificate.get_by_id(certificate_id)
     if cert.user != current_user and not current_user.is_admin():
         abort(403)
@@ -85,6 +100,10 @@ def download(certificate_id, download_type):
 @certificate.route('/<int:certificate_id>/revoke')
 @login_required
 def revoke(certificate_id):
+    use_ssl = Setting.get_by_name('use_ssl', default=False).value
+    if use_ssl == False:
+        abort(404)
+
     cert = Certificate.get_by_id(certificate_id)
     if cert.user != current_user and not current_user.is_admin():
         abort(403)
@@ -104,6 +123,10 @@ def revoke(certificate_id):
 @login_required
 @admin_required
 def generateCA():
+    use_ssl = Setting.get_by_name('use_ssl', default=False).value
+    if use_ssl == False:
+        abort(404)
+
     ca_cert = Certificate(
                 name="AlarmDecoder CA",
                 description='CA certificate used for authenticating others.',
@@ -175,6 +198,10 @@ def _update_ser2sock_config(path):
 @login_required
 @admin_required
 def revokeCA():
+    use_ssl = Setting.get_by_name('use_ssl', default=False).value
+    if use_ssl == False:
+        abort(404)
+
     ca = Certificate.query.filter_by(type=CA).first()
     certs = Certificate.query.filter_by(ca_id=ca.id).delete()
     ca = Certificate.query.filter_by(type=CA).delete()
